Factor VMTest out of IsolateTest
This code will be shared with SubprocessTest to
support running tests in another process.
